Time: 0 ms (100%), Space: 113.8 MB (44.96%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.cpp
@@ -1,11 +1,11 @@
 class Solution {
 public:
     double findMaxAverage(vector<int>& nums, int k) {
-        long long sum = 0;
+        int sum = 0;
         for(int i = 0; i < k; i++) {
             sum += nums[i];
         }
-        long long bestSum = sum;
+        int bestSum = sum;
         for(int i = k; i < nums.size(); i++) {
             sum -= nums[i - k];
             sum += nums[i];
